pyproject.toml
[tool.poetry]
name = "test-stripe"
version = "0.1.0"
description = ""
authors = ['Kimiyori']
readme = "README.md"
packages = [{include = "test_stripe"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
psycopg2-binary = "^2.9.5"
stripe = "^5.0.0"
djangorestframework = "^3.14.0"
environs = "^9.5.0"
dependency-injector = "^4.40.0"
django-debug-toolbar = "^3.7.0"

[tool.poetry.group.linters.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
pylint = "^2.15.6"
mypy = "0.982"
django-stubs = {extras = ["compatible-mypy"], version = "^1.13.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
